Moves OrthoProjection to camera package
