+++ src/pyrite/_camera/ortho_projection.py
