+++ src/pyrite/camera.py
@@ -1,5 +1,7 @@
 import pyrite._camera.camera
 import pyrite._camera.chase_camera
+import pyrite._camera.ortho_projection
 
 Camera = pyrite._camera.camera.BaseCamera
 ChaseCamera = pyrite._camera.chase_camera.ChaseCamera
+OrthoProjection = pyrite._camera.ortho_projection.OrthoProjection
