+++ src/pyrite/rendering.py
@@ -1,6 +1,5 @@
 import pyrite._rendering.base_renderable
 import pyrite._rendering.camera_renderer
-import pyrite._rendering.ortho_projection
 import pyrite._rendering.rect_bounds
 import pyrite._rendering.render_texture
 import pyrite._component.render_texture_component
@@ -9,7 +8,6 @@ import pyrite._rendering.viewport
 
 Renderable = pyrite._rendering.base_renderable.BaseRenderable
 CameraRenderer = pyrite._rendering.camera_renderer.CameraRendererProvider
-OrthoProjection = pyrite._rendering.ortho_projection.OrthoProjection
 RectBounds = pyrite._rendering.rect_bounds.RectBounds
 RenderTexture = pyrite._rendering.render_texture.RenderTexture
 RenderTextureComponent = (
